analytics_engine_1/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv,asyncio]
pandas
python-dotenv